Add support to export logs to syslog
It's a requirement of some systems to push logs to a central aggregation
tool, such that they can then be exported off the machine to a central
store.

Syslog is the common system to do this, or where syslog itself is not
used common logging tools (such as systemd-journald) implement a syslog
compatible interface. Additionally, vsftpd supports syslog by default
with it's "syslog_enable" option, however this is not yet implemented as
configuration in this repository.

This commit implements this setting, with a default to "NO" -- the
current behaviour.