Modify the way the users are managed
Currently, users are added via Ansible, and can then be manually
modified through the user of a script. However, the author believes that
this approach has certain negative side effects; namely, it's possible
to create "out of band" management of users through the use of this
script.

While it's possible for Ansible to manage the users created and managed
by this script, the author cannot reconcile this change with the more
general approach of infrastructure as code.

Accordingly, this commit replaces user management entirely with Ansible
primitives; in particular, the htpasswd module. This allows both adding
and deleting users through variable declarations.

Design notes:

no folder deletion:

  Currently, as part of the creation of users folders are also created
  in the vsftpd root. However, they are not removed when users are also
  removed.

  This follows the pattern already established by the use of the
  user script. Future work should consider whether the data of these
  dirs should be removed when the users are removed.

/etc/vsftpd now 600

  Because /etc/vsftpd is now being managed by Ansible, permissions are
  specified. The new permissions are less permissive than they were
  previously, as vsftpd runs as root and can read this file regardless.

  To make this clear, the file is 600, or accessible only by the root
  user (who can access this file regardless)

sha512 algorythm

  sha512 was chosen as the password algorythm as it's supported by
  the crypt_r system calls used by the pam module[1][2], and htpasswd
  passlib module.

[1] - https://packages.debian.org/source/jessie/glibc)
[2] - https://serverfault.com/questions/330069/how-to-create-an-sha-512-hashed-password-for-shadow